Finished
Project instatiates a navcontroller as the root view, then uses a viewcontroller to instantiate about 10 rows of custom views. Each view contains an editable label, a stepper, and a numerical indicator of the count from the stepper. In this version, the keyboard does not automatically show when the label is selected, but it hides when text entry is finished.